Replace Assembler_PropertiesSpec with Assembler_PropertiesTests.
